backend wallets, transactions and reviews tables

+++ BackEnd/database/migrations/2024_05_14_035646_create_jobs_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
             $table->id();
             $table->string('title');
             $table->text('description');
-            $table->float('salary');
+            $table->double('salary');
             $table->enum('type', ['PartTime', 'FullTime']);
             $table->boolean('is_done');
             $table->foreignId('company_id')->constrained('companies');
